# workshop/services/sample/src/generated/schemas.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export const ResourceStatusSchema = {
    type: 'string',
    description: 'Status of a resource',
    enum: ['active', 'inactive', 'pending', 'archived']
} as const;

export const ResourceSchema = {
    type: 'object',
    properties: {
        id: {
            type: 'string',
            description: 'Unique identifier for the resource'
        },
        name: {
            type: 'string',
            description: 'Name of the resource'
        },
        description: {
            type: 'string',
            description: 'Detailed description of the resource'
        },
        status: {
            '$ref': '#/components/schemas/ResourceStatus',
            description: 'Current status of the resource'
        },
        tags: {
            type: 'array',
            items: {
                type: 'string'
            },
            description: 'Tags associated with the resource'
        },
        createdAt: {
            type: 'string',
            format: 'date-time',
            description: 'When the resource was created'
        },
        updatedAt: {
            type: 'string',
            format: 'date-time',
            description: 'When the resource was last updated'
        }
    },
    required: ['id', 'name', 'status', 'createdAt', 'updatedAt']
} as const;

export const ResourceCreateSchema = {
    type: 'object',
    properties: {
        name: {
            type: 'string',
            description: 'Name of the resource'
        },
        description: {
            type: 'string',
            description: 'Detailed description of the resource'
        },
        status: {
            '$ref': '#/components/schemas/ResourceStatus',
            description: 'Initial status of the resource',
            default: 'active'
        },
        tags: {
            type: 'array',
            items: {
                type: 'string'
            },
            description: 'Tags associated with the resource'
        }
    },
    required: ['name']
} as const;

export const ResourceUpdateSchema = {
    type: 'object',
    properties: {
        name: {
            type: 'string',
            description: 'Updated name of the resource'
        },
        description: {
            type: 'string',
            description: 'Updated description of the resource'
        },
        status: {
            '$ref': '#/components/schemas/ResourceStatus',
            description: 'Updated status of the resource'
        },
        tags: {
            type: 'array',
            items: {
                type: 'string'
            },
            description: 'Updated tags for the resource'
        }
    }
} as const;

export const ResourceListSchema = {
    type: 'object',
    properties: {
        items: {
            type: 'array',
            items: {
                '$ref': '#/components/schemas/Resource'
            },
            description: 'List of resources'
        },
        total: {
            type: 'number',
            description: 'Total number of resources matching the query'
        },
        page: {
            type: 'number',
            description: 'Current page number'
        },
        page_size: {
            type: 'number',
            description: 'Number of items per page'
        }
    },
    required: ['items', 'total', 'page', 'page_size']
} as const;

export const ErrorSchema = {
    type: 'object',
    properties: {
        error: {
            type: 'object',
            properties: {
                code: {
                    type: 'string',
                    description: 'Error code'
                },
                message: {
                    type: 'string',
                    description: 'Human-readable error message'
                },
                details: {
                    type: 'object',
                    additionalProperties: true,
                    description: 'Additional error details'
                }
            },
            required: ['code', 'message']
        }
    },
    required: ['error']
} as const;